FIX compare baseline against current code NOT most recent commit.
Previously SARB generated a diff from the baseline commit and the most
recent commit. However static anlaysis might be run on the current
state of the code. These 2 might be different.

When running on CI the current state of the code and latest commit will
_probably_ be the same.

Developing locally this will probably not be the case. Developers will
want to make sure no errors have been introduced since the baseline
before committing code. To improve DX SARB will look at diff between the
baseline commit and HEAD.

It is assumed the process will be something like this:

- developer edits code
- developer runs static analysis tool
- developer runs SARB
- developer removes any issues raised since the baseline
- once all post baseline issues are removed developer commits code

Hence the need for the diff to be taken against current state of code
rather than last commit.